fix for Ctrl K bug with no doc open
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@1632 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/scribus.cpp
@@ -7453,8 +7453,9 @@ void ScribusApp::MakeFrame(int f, int c, double *vals)
 
 void ScribusApp::DeleteObjekt()
 {
-	if (!doc->EditClip)
-		view->DeleteItem();
+	 if (HaveDoc)
+	 	if (!doc->EditClip)
+			view->DeleteItem();
 }
 
 void ScribusApp::Objekt2Back()
